requirements.txt
# MBTA Dashboard Requirements
# Original dependencies
pandas
requests
vaderSentiment

# NEW: Google Sheets API dependencies (for live editing feature)
google-auth>=2.0.0
google-auth-oauthlib>=0.5.0
google-auth-httplib2>=0.1.0
google-api-python-client>=2.0.0

# Optional but recommended for production:
# python-dotenv  # For managing environment variables from .env file
